+++ .buildkite/ftr_oblt_stateful_configs.yml
@@ -41,7 +41,6 @@ enabled:
   - x-pack/test/observability_functional/with_rac_write.config.ts
   - x-pack/test/observability_onboarding_api_integration/basic/config.ts
   - x-pack/test/observability_onboarding_api_integration/cloud/config.ts
-  - x-pack/test/observability_ai_assistant_api_integration/enterprise/config.ts
   - x-pack/test/observability_ai_assistant_functional/enterprise/config.ts
   - x-pack/solutions/observability/test/api_integration/profiling/cloud/config.ts
   - x-pack/test/functional/apps/apm/config.ts
