+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/ai_assistant/complete/functions/alerts.spec.ts
@@ -11,10 +11,7 @@ import { InternalRequestHeader, RoleCredentials } from '@kbn/ftr-common-function
 import { ApmSynthtraceEsClient } from '@kbn/apm-synthtrace';
 import { ApmRuleType } from '@kbn/rule-data-utils';
 import { SearchAlertsResult } from '@kbn/alerts-ui-shared/src/common/apis/search_alerts/search_alerts';
-import {
-  LlmProxy,
-  createLlmProxy,
-} from '../../../../../../../observability_ai_assistant_api_integration/common/create_llm_proxy';
+import { LlmProxy, createLlmProxy } from '../../utils/create_llm_proxy';
 import {
   getMessageAddedEvents,
   invokeChatCompleteWithFunctionRequest,
