+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/ai_assistant/utils/create_openai_chunk.ts
