+++ x-pack/test/functional/apps/search_playground/playground_overview.ess.ts
@@ -12,7 +12,7 @@ import { MachineLearningCommonAPIProvider } from '../../services/ml/common_api';
 import {
   createLlmProxy,
   LlmProxy,
-} from '../../../observability_ai_assistant_api_integration/common/create_llm_proxy';
+} from '../../../api_integration/deployment_agnostic/apis/observability/ai_assistant/utils/create_llm_proxy';
 
 const esArchiveIndex =
   'src/platform/test/api_integration/fixtures/es_archiver/index_patterns/basic_index';
