+++ x-pack/test/functional/apps/search_playground/utils/create_openai_connector.ts
@@ -6,7 +6,7 @@
  */
 
 import type SuperTest from 'supertest';
-import { LlmProxy } from '../../../../observability_ai_assistant_api_integration/common/create_llm_proxy';
+import { LlmProxy } from '../../../../api_integration/deployment_agnostic/apis/observability/ai_assistant/utils/create_llm_proxy';
 
 export async function createOpenAIConnector({
   supertest,
