--- x-pack/test/observability_ai_assistant_api_integration/enterprise/config.ts
@@ -1,11 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { configs } from '../configs';
-
-// eslint-disable-next-line import/no-default-export
-export default configs.enterprise;
