+++ x-pack/test/observability_ai_assistant_functional/common/create_synthtrace_client.ts
@@ -10,7 +10,7 @@ import {
   createLogger,
   LogLevel,
 } from '@kbn/apm-synthtrace';
-import { InheritedFtrProviderContext } from './ftr_provider_context';
+import { InheritedFtrProviderContext } from '../ftr_provider_context';
 
 export async function getApmSynthtraceEsClient(
   context: InheritedFtrProviderContext,
