+++ x-pack/test/observability_ai_assistant_functional/common/ui/index.ts
@@ -9,7 +9,7 @@ import type { PathsOf, TypeAsArgs, TypeOf } from '@kbn/typed-react-router-config
 import { kbnTestConfig } from '@kbn/test';
 import type { ObservabilityAIAssistantRoutes } from '@kbn/observability-ai-assistant-app-plugin/public/routes/config';
 import qs from 'query-string';
-import { User } from '../../../observability_ai_assistant_api_integration/common/users/users';
+import { User } from '../users/users';
 import type { InheritedFtrProviderContext } from '../../ftr_provider_context';
 
 export interface ObservabilityAIAssistantUIService {
