+++ x-pack/test/observability_ai_assistant_functional/common/users/create_users_and_roles.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { InheritedFtrProviderContext } from '../ftr_provider_context';
+import { InheritedFtrProviderContext } from '../../ftr_provider_context';
 import { allUsers } from './users';
 import { allRoles } from './roles';
