+++ x-pack/test/observability_ai_assistant_functional/tests/conversations/archiving.spec.ts
@@ -9,7 +9,7 @@ import expect from '@kbn/expect';
 import {
   createLlmProxy,
   LlmProxy,
-} from '../../../observability_ai_assistant_api_integration/common/create_llm_proxy';
+} from '../../../api_integration/deployment_agnostic/apis/observability/ai_assistant/utils/create_llm_proxy';
 import { FtrProviderContext } from '../../ftr_provider_context';
 import { createConnector, deleteConnectors } from '../../common/connectors';
 import { deleteConversations } from '../../common/conversations';
