+++ x-pack/test/observability_ai_assistant_functional/tests/feature_controls/assistant_security.spec.ts
@@ -10,7 +10,7 @@ import { FtrProviderContext } from '../../ftr_provider_context';
 import {
   createLlmProxy,
   LlmProxy,
-} from '../../../observability_ai_assistant_api_integration/common/create_llm_proxy';
+} from '../../../api_integration/deployment_agnostic/apis/observability/ai_assistant/utils/create_llm_proxy';
 import { createConnector, deleteConnectors } from '../../common/connectors';
 import { createAndLoginUserWithCustomRole, deleteAndLogoutUser } from './helpers';
