+++ x-pack/test/observability_ai_assistant_functional/tests/knowledge_base/index.spec.ts
@@ -10,7 +10,7 @@ import { KnowledgeBaseState } from '@kbn/observability-ai-assistant-plugin/commo
 import {
   LlmProxy,
   createLlmProxy,
-} from '../../../observability_ai_assistant_api_integration/common/create_llm_proxy';
+} from '../../../api_integration/deployment_agnostic/apis/observability/ai_assistant/utils/create_llm_proxy';
 import { FtrProviderContext } from '../../ftr_provider_context';
 import {
   deployTinyElserAndSetupKb,
