+++ x-pack/test_serverless/tsconfig.json
@@ -84,7 +84,6 @@
     "@kbn/search-types",
     "@kbn/config-schema",
     "@kbn/features-plugin",
-    "@kbn/observability-ai-assistant-plugin",
     "@kbn/test-suites-src",
     "@kbn/console-plugin",
     "@kbn/cloud-security-posture-common",
